+++ opencascade-hs/src/OpenCascade/Poly/Types.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.Poly.Types 
+(Triangulation
+, Triangle
+) where
+
+data Triangulation
+data Triangle